feat: add DAR comprovante generation service
